Fixing Tests
